+++ code/App_Client/lib/view/cart/widget/widget_bottomNavigationBar.dart
@@ -4,7 +4,10 @@ import 'package:app_pizza_client/widget/custom/costum_bar.dart';
 import 'package:app_pizza_client/widget/custom/costum_botton.dart';
 import 'package:flutter/material.dart';
 
-Widget widget_BottomNavigationBar(BuildContext context,{required double priceTotal}) {
+Widget widget_BottomNavigationBar(
+  BuildContext context, {
+  required int priceTotal,
+}) {
   return Container(
     height: context.heightPct(20),
     width: double.infinity,
@@ -55,7 +58,6 @@ Widget widget_BottomNavigationBar(BuildContext context,{required double priceTot
             width: 90,
           ),
         ),
-        
       ],
     ),
   );
